initial refactoring;
