+++ .nvmrc
@@ -1,0 +1,1 @@
+v14.16.0
