+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "sum.js",
   "scripts": {
+    "format": "prettier --write \"src/**/*.js\" \"test/**/*.js\"",
     "test": "jest",
     "test-update": "jest --updateSnapshot",
     "start": "node src"
@@ -22,6 +23,7 @@
     "lodash": "^4.17.21",
     "mongodb": "^3.6.6",
     "mongodb-memory-server": "^6.9.6",
-    "morgan": "^1.10.0"
+    "morgan": "^1.10.0",
+    "prettier": "^2.2.1"
   }
 }
