+++ src/database/ads.js
@@ -1,12 +1,14 @@
 // ./src/database/ads.js
-const {getDatabase} = require('./mongo');
-const {ObjectID} = require('mongodb');
+const { getDatabase } = require('./mongo');
+const { ObjectID } = require('mongodb');
 
 const collectionName = 'ads';
 
 async function insertAd(ad) {
   const database = await getDatabase();
-  const {insertedId} = await database.collection(collectionName).insertOne(ad);
+  const { insertedId } = await database
+    .collection(collectionName)
+    .insertOne(ad);
   return insertedId;
 }
 
@@ -16,24 +18,24 @@ async function getAds() {
 }
 
 async function deleteAd(id) {
-    const database = await getDatabase();
-    await database.collection(collectionName).deleteOne({
-      _id: new ObjectID(id),
-    });
-  }
-  
-  async function updateAd(id, ad) {
-    const database = await getDatabase();
-    delete ad._id;
-    await database.collection(collectionName).update(
-      { _id: new ObjectID(id), },
-      {
-        $set: {
-          ...ad,
-        },
+  const database = await getDatabase();
+  await database.collection(collectionName).deleteOne({
+    _id: new ObjectID(id),
+  });
+}
+
+async function updateAd(id, ad) {
+  const database = await getDatabase();
+  delete ad._id;
+  await database.collection(collectionName).update(
+    { _id: new ObjectID(id) },
+    {
+      $set: {
+        ...ad,
       },
-    );
-  }
+    },
+  );
+}
 
 module.exports = {
   insertAd,
