+++ src/database/mongo.js
@@ -1,15 +1,17 @@
 // ./src/database/mongo.js
-const {MongoMemoryServer} = require('mongodb-memory-server');
-const {MongoClient} = require('mongodb');
+const { MongoMemoryServer } = require('mongodb-memory-server');
+const { MongoClient } = require('mongodb');
 
 let database = null;
 
 async function startDatabase() {
   const mongo = new MongoMemoryServer();
   const dbPath = await mongo.getDbPath();
-  console.log(dbPath)
+  console.log(dbPath);
   const mongoDBURL = await mongo.getConnectionString();
-  const connection = await MongoClient.connect(mongoDBURL, {useNewUrlParser: true});
+  const connection = await MongoClient.connect(mongoDBURL, {
+    useNewUrlParser: true,
+  });
   database = connection.db();
 }
