+++ src/index.js
@@ -1,6 +1,6 @@
-const server = require('./server')
+const server = require('./server');
 
 server().then(async () => {
-  await insertAd({title: 'Hello, now from the in-memory database!'});
-  console.log("Server is running")
-})
+  await insertAd({ title: 'Hello, now from the in-memory database!' });
+  console.log('Server is running');
+});
