--- sum.test.js
@@ -1,5 +1,0 @@
-const sum = require('./sum')
-
-test('add 1 + 2 to equals 3', () => {
-    expect(sum(1,2)).toBe(3)
-})
